Rename: Manager를 기능에 맞게 패키지를 분리한다.
